% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rollbar.R
\name{rollbar.critical}
\alias{rollbar.critical}
\title{Report critical errors}
\usage{
rollbar.critical(message, extra = NULL)
}
\arguments{
\item{message}{The message or error}

\item{extra}{Extra data}
}
\description{
Report critical errors
}
