% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rollbar.R
\name{rollbar.debug}
\alias{rollbar.debug}
\title{Report debug messages}
\usage{
rollbar.debug(message, extra = NULL)
}
\arguments{
\item{message}{The message or error}

\item{extra}{Extra data}
}
\description{
Report debug messages
}
