% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rollbar.R
\name{rollbar.error}
\alias{rollbar.error}
\title{Report errors}
\usage{
rollbar.error(message, extra = NULL)
}
\arguments{
\item{message}{The message or error}

\item{extra}{Extra data}
}
\description{
Report errors
}
