% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rollbar.R
\name{rollbar.info}
\alias{rollbar.info}
\title{Report info messages}
\usage{
rollbar.info(message, extra = NULL)
}
\arguments{
\item{message}{The message or error}

\item{extra}{Extra data}
}
\description{
Report info messages
}
\examples{
rollbar.info("Job successful")
rollbar.info("Job successful", list(job_id=123, awesome="yes"))
}
