% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rollbar.R
\name{rollbar.log}
\alias{rollbar.log}
\title{Report messages}
\usage{
rollbar.log(level, message, extra = NULL)
}
\arguments{
\item{level}{The "error" level}

\item{message}{The message or error}

\item{extra}{Extra data}
}
\description{
Report messages
}
\examples{
rollbar.log("info", "Job successful")
rollbar.log("info", "Job successful", list(job_id=123, awesome="yes"))
}
