% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rollbar.R
\name{rollbar.warning}
\alias{rollbar.warning}
\title{Report warnings}
\usage{
rollbar.warning(message, extra = NULL)
}
\arguments{
\item{message}{The message or error}

\item{extra}{Extra data}
}
\description{
Report warnings
}
